migrate to vulkan renderer build ok, but not work now

+++ include/ShaderUtils.hpp
@@ -5,8 +5,6 @@
 #ifndef DRENDER_SHADERUTILS_HPP
 #define DRENDER_SHADERUTILS_HPP
 
-#include "glad/glad.h"
-
 #include <string>
 
 class ShaderUtils final
@@ -20,8 +18,8 @@ public:
 		FRAGMENT, VERTEX
 	};
 
-	static GLuint LoadShader(std::string filepath, ShaderType type);
-	static GLuint LoadShaders(std::string vertex, std::string fragment);
+	static size_t LoadShader(std::string filepath, ShaderType type);
+	static size_t LoadShaders(std::string vertex, std::string fragment);
 
 private:
 	ShaderUtils(const ShaderUtils&) = delete;
