+++ src/Scene/SceneManager.cpp
@@ -1,8 +1,12 @@
 //
 // Created by debyecao on 11/25/20.
 //
+#include "stdafx.hpp"
+
 #include "Scene/SceneManager.hpp"
+#include "Scene/BaseScene.hpp"
 #include "Logger.hpp"
+#include "GameContext.hpp"
 
 #include <algorithm>
 
@@ -56,7 +60,7 @@ void SceneManager::RemoveScene(BaseScene* scene)
 
 void SceneManager::SetCurrentScene(int sceneIndex)
 {
-	if (sceneIndex < 0 || sceneIndex >= m_Scenes.size())
+	if (sceneIndex < 0 || sceneIndex >= (int)m_Scenes.size())
 	{
 		Logger::LogError("Could not set scene to index " + std::to_string(sceneIndex) +
 		                 ", it doesn't exist in the scene manager!");
@@ -81,7 +85,7 @@ void SceneManager::SetCurrentScene(std::string sceneName)
 
 BaseScene* SceneManager::CurrentScene() const
 {
-	if (m_CurrentSceneIndex < 0 || m_CurrentSceneIndex >= m_Scenes.size())
+	if (m_CurrentSceneIndex < 0 || m_CurrentSceneIndex >= (int)m_Scenes.size())
 	{
 		return nullptr;
 	}
